entries/urls.py
# entries/urls.py
from rest_framework.routers import DefaultRouter
from .views import EntryViewSet

router = DefaultRouter()
router.register(r"entries", EntryViewSet, basename="entry")

urlpatterns = router.urls